requirements.txt
functions-framework>=3.0.0
firebase-functions>=1.0.0
requests>=2.31.0
Flask>=3.0.0
webauthn>=2.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=21.2.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
setuptools>=65.0.0
google-cloud-scheduler>=2.0.0
